% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_descriptors.R
\name{create.descriptors}
\alias{create.descriptors}
\title{Create rdkit descriptors}
\usage{
create.descriptors(file_name, descriptors = NULL)
}
\arguments{
\item{descriptors}{A vector containing the desired descriptors. This argument is optional. If not
provided, then the function returns the complete list of generated descriptors}

\item{file}{A .xlsx file, with two columns: a 'Name' column and a 'SMILES' column, containing
the molecule names and SMILES representations, respectively}
}
\value{
A vector of rcdk descriptors
}
\description{
Based on a .xlsx file of specific structure and a subset of rdkit descriptors, this function 
returns the values of the specific descriptors for all requested molecules
}
\details{
The user can generate rcdk descriptors using the
 function \code{create.descriptors()}. Additionally, a subset of the complete rcdk descriptor
 list can be returned leveraging the argument 'descriptors'. Note that in order for this function to
 be excecuted, the following libraries need to be installed: rJava, rlang, Rcpp
 openxlsx, rsvg, rcdk, ChemmineR, zlibbioc, BiocGenerics and ChemmineOB
}
\examples{
 \dontrun{
 #tree.model <- create.descriptors("SMILES.xlsx", descriptors = c("VP.7", "MDEC.34","VPC.4","fragC"))
}

}
