% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_bn.R
\name{deploy.bn}
\alias{deploy.bn}
\title{Deploy Bayesian Netowrks  on Jaqpot}
\usage{
deploy.bn(
  trained.model,
  implemented.in = "bnlearn",
  type = "discrete",
  predict.args = list(method = "bayes-lw", n = 1000),
  url = "https://api.jaqpot.org/jaqpot/",
  ...
)
}
\arguments{
\item{trained.model}{An object of class ""bn.fit"" (function \code{bn.fit(()} of package 'bnlearn').}

\item{implemented.in}{The library the model was build on. Currently, only "bnlearn" is supported.}

\item{type}{The type of the bayesian network, with two options: "discrete" or "gaussian"}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}

\item{...}{Extra arguments to be passed down the R client. This is not recommended.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained bayesian networks on Jaqpot
}
\details{
The user can upload on Jaqpot a model that has been trained using the
 function \code{bn.fit()} of package 'bnlearn'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #my_model <- bnlearn::bn.fit(x = structure, x = data, method="bayes")
 #deploy.bn(my_model)
}

}
