% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_caret.R
\name{deploy.caret}
\alias{deploy.caret}
\title{Deploy Decision Tree Models on Jaqpot}
\usage{
deploy.caret(
  preprocess.model = NULL,
  trained.model,
  url = "https://api.jaqpot.org/jaqpot/",
  ...
)
}
\arguments{
\item{preprocess.model}{The preprocess model. Here it is very important not to include the response into the preprocessing phase!
Allowed types of preprocessing: 'preProcess' and 'dummyVars' caret functions.}

\item{trained.model}{An object of class "tree" (function \code{ctree(()} of package 'party').}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.
#' @param ... Extra arguments to be passed down the R client. This is not recommended.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained decision tree models on Jaqpot given
a "tree" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the
 function \code{ctree()} of package 'party'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #party.model <- party::ctree(y~x, data=df)
 #deploy.party(tree.model)
}

}
