% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_lm_glm.R
\name{deploy.lm}
\alias{deploy.lm}
\title{Deploy (Generalized) Linear Models on Jaqpot}
\usage{
deploy.lm(object, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{object}{An object of class "lm" (base function \code{lm()}) or "glm"
(base function \code{glm()}).}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained linear and generalized linear models on Jaqpot given
an "lm" or "glm" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the base
 functions \code{lm()} or \code{glm()}. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #lm.model <- lm(y~x, data=df)
 #deploy.lm(lm.model)

 #glm.model <- glm(y~x, data=df, family =  "gaussian")
 #deploy.lm(glm.model)
 }

}
