% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_naive_bayess.R
\name{deploy.naive.bayess}
\alias{deploy.naive.bayess}
\title{Deploy Naive Bayess models on Jaqpot}
\usage{
deploy.naive.bayess(object)
}
\arguments{
\item{object}{An object of either class "naive_bayess" (base function \code{naiveBayess()})
(base function \code{naiveBayess()})}
}
\value{
The id of the uploaded model
}
\description{
Uploads trained linear and generalized linear models on Jaqpot given
an "naive_bayess" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the base
 function \code{naiveBayess()}. The data used for training are deleted before the
 model is uploaded on the platform. Apart from the model object, the user is requested
 to provide further information (e.g. Jaqpot API key or credentials, model title, short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 #nb.model <- naivebayes::naive_bayes(y~x, data=df)
 #deploy.naive.bayess(nb.model)


}
