% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_naive_bayess.R
\name{deploy.nb}
\alias{deploy.nb}
\title{Deploy Naive Bayes (NB) Models on Jaqpot}
\usage{
deploy.nb(object, dependent.vars, url = "https://api.jaqpot.org/")
}
\arguments{
\item{object}{An object of class "naive_bayes"  (function \code{naive_bayes()} of package 'naivebayes').}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained NB models on Jaqpot given a "naive_bayes" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the 
 function \code{naive_bayes()} of package 'naivebayes'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #nb.model <- naivebayes::naive_bayes(y~x, data=df)
 #deploy.nb(nb.model)
 }


}
