% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_party.R
\name{deploy.party}
\alias{deploy.party}
\title{Deploy (party) Tree models on Jaqpot}
\usage{
deploy.party(object)
}
\arguments{
\item{object}{An object of  (party function \code{ctree(()}) or "party model"
(party function \code{ctree(()})}
}
\value{
The id of the uploaded model
}
\description{
Uploads trained tree tree regression model on Jaqpot given
a "tree" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the base
 function \code{ctree()}. The data used for training are deleted before the
 model is uploaded on the platform. Apart from the model object, the user is requested
 to provide further information (e.g. Jaqpot API key or credentials, model title, short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 #party.model <- party(y~x, data=df)
 #deploy.tree(tree.model)


}
