% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_rpart.R
\name{deploy.rpart}
\alias{deploy.rpart}
\title{Deploy  Classification  and Regression Tree Models on Jaqpot}
\usage{
deploy.rpart(object, replace = NULL, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{object}{An object of class "rpart" (function \code{rpart()} of package 'rpart').}

\item{replace}{used for NA substitution with a desired numeric value.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained rpart tree regression and classification models on Jaqpot given
an "rpart" object.
}
\details{
The user can upload on Jaqpot an rpart model that has been trained using the
 function \code{rpart()} of package 'rpart'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #rpart.model <- rpart::rpart(y~x, data=df)
 #deploy.rpart(rpart.model)
}

}
