% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_rpart.R
\name{deploy.rpart.regr.tree}
\alias{deploy.rpart.regr.tree}
\title{Deploy (rpart) Regression tree models on Jaqpot}
\usage{
deploy.rpart.regr.tree(object)
}
\arguments{
\item{object}{An object of  (base function \code{rpart()}) or "rpart model"
(base function \code{rpart()})}
}
\value{
The id of the uploaded model
}
\description{
Uploads trained rpart tree model on Jaqpot given
a "tree" object.
}
\details{
The user can upload on Jaqpot a classification rpart model that has been trained using the base
 function \code{rpart()} . The data used for training are deleted before the
 model is uploaded on the platform. Apart from the model object, the user is requested
 to provide further information (e.g. Jaqpot API key or credentials, model title, short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 #rpart.model <- rpart(y~x, data=df)
 #deploy.rpart.tree(rpart.model)


}
