% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.dataset.byID}
\alias{get.dataset.byID}
\title{Get the Information of a Jaqpot Dataset}
\usage{
get.dataset.byID(datasetID, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{datasetID}{A string containing the Jaqpot id of the dataset.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
All information relevant to the dataset.
}
\description{
The user can obtain the Information of a Jaqpot dataset given the dataset id.
}
\details{
The user can obtain the information of a dataset that has been uploaded on Jaqpot by providing 
the appropriate dataset id.Notethat the user should also be the dataset creator.
}
\examples{
 \dontrun{
#dataset_info <- get.dataset.byID(datasetID = "Lfz3aBdh4LlJIJxAvNqV")
}

}
