% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.dataset.feats}
\alias{get.dataset.feats}
\title{Get the Features of a Jaqpot dataset}
\usage{
get.dataset.feats(datasetID, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{datasetID}{A string containing the Jaqpot id of the dataset.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
A vector containig the features of the dataset.
}
\description{
The user can obtain the features of a Jaqpot dataset given the dataset id.
}
\details{
The user can obtain the features of a dataset that has been uploaded on Jaqpot by providing 
the appropriate dataset id. Note that that the user should also be the dataset creator.
}
\examples{
 \dontrun{
#dataset_feats <- get.dataset.feats(datasetID = "Lfz3aBdh4LlJIJxAvNqV")
}

}
