% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.feature.byID}
\alias{get.feature.byID}
\title{Get the Information of a Jaqpot Feature}
\usage{
get.feature.byID(featID, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{featID}{A string containing the Jaqpot id of the feature.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
All information relevant to the feature.
}
\description{
The user can obtain the Information of a Jaqpot feature given the feature id.
}
\details{
The user can obtain the information of a feature that is related to a dataset/model that has been uploaded on Jaqpot by providing 
the appropriate feature id.Notethat the user should also be the feature creator.
}
\examples{
 \dontrun{
#feature_info <- get.feature.byID(featID = "Lfz3aBdh4LlJIJxAvNqV")
}

}
