% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.model.feats}
\alias{get.model.feats}
\title{Get the Independent Features of a Jaqpot Model}
\usage{
get.model.feats(modelID, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{modelID}{A string containing the Jaqpot id of the model.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
A vector containig the independent features of the model.
}
\description{
The user can obtain the independent features of a Jaqpot Model given the model id.
}
\details{
The user can obtain the independent features of a model that has been uploaded on Jaqpot by providing 
the appropriate model id. Note that that the user should also be the model creator.
}
\examples{
 \dontrun{
#model_feats <- get.model.feats(modelID = "Lfz3aBdh4LlJIJxAvNqV")
}

}
