% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.models.byTag}
\alias{get.models.byTag}
\title{Get a Number of Models of a Specific Tag}
\usage{
get.models.byTag(
  tag,
  min = 0,
  max = 10,
  url = "https://api.jaqpot.org/jaqpot/"
)
}
\arguments{
\item{tag}{A string containing the model tag.}

\item{min}{The minimum number of models to return.}

\item{max}{The maximum number of models to return.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
All information of the upoaded models that contain the selected tag.
}
\description{
The user can obtain a number of models that have been uploaded on Jaqpot and contain a specific tag.
}
\details{
The user can obtain the information of a number of models that have been uploaded on Jaqpot and contain a specific tag.
Notethat the user should also be the model creator.
}
\examples{
 \dontrun{
#get_tag_models <- get.tag.models()
}

}
