% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.my.datasets}
\alias{get.my.datasets}
\title{Get a Number of Datasets Uploaded On Jaqpot}
\usage{
get.my.datasets(min = 0, max = 10, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{max}{The maximum number of datasets to return.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
All information of the upoaded datasets
}
\description{
The user can obtain a number of datasets uploaded on Jaqpot.
 
 @param min The minimum number of datasets to return.
}
\details{
The user can obtain the information of a number of datasets that have been uploaded on Jaqpot.
Notethat the user should also be the dataset creator for all models.
}
\examples{
 \dontrun{
#get_datasets <- get.my.datasets()
}

}
