% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.my.models}
\alias{get.my.models}
\title{Get a Number of Models Uploaded On Jaqpot}
\usage{
get.my.models(min = 0, max = 10, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{max}{The maximum number of models to return.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
All information of the upoaded models.
}
\description{
The user can obtain a number of models uploaded on Jaqpot.
 
 @param min The minimum number of models to return.
}
\details{
The user can obtain the information of a number of models that have been uploaded on Jaqpot.
Notethat the user should also be the model creator for all models.
}
\examples{
 \dontrun{
#get_models <- get.my.models()
}

}
