% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.orgs.models.byTag}
\alias{get.orgs.models.byTag}
\title{Get a Number of Models of a Specific Tag that Belong to a Specific Organisation}
\usage{
get.orgs.models.byTag(
  orgID,
  tag,
  min = 0,
  max = 10,
  url = "https://api.jaqpot.org/jaqpot/"
)
}
\arguments{
\item{orgID}{A string containing the Jaqpot id of the organisation.}

\item{tag}{A string containing the model tag.}

\item{min}{The minimum number of models to return.}

\item{max}{The maximum number of models to return.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
All information of the upoaded models that contain the selected tag and belong to the selected organisation.
}
\description{
The user can obtain a number of models that have been uploaded on Jaqpot, contain a specific tag and belong to a certain organisation.
}
\details{
The user can obtain the information of a number of models that have been uploaded on Jaqpot, contain a specific tag and belong to a specific organisation.
Notethat the user should be part of the organisation .
}
\examples{
 \dontrun{
#get_tagandOrg_models <- get.orgs.models.byTag()
}

}
