% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.orgs.models}
\alias{get.orgs.models}
\title{Get a Number of Models Belonging to a Certain Organisation}
\usage{
get.orgs.models(
  orgID,
  min = 0,
  max = 10,
  url = "https://api.jaqpot.org/jaqpot/"
)
}
\arguments{
\item{orgID}{A string containing the Jaqpot id of the organisation.}

\item{min}{The minimum number of models to return.}

\item{max}{The maximum number of models to return.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
All information of the upoaded models of the selected organisation.
}
\description{
The user can obtain a number of models that have been uploaded on Jaqpot and belong to a specific Organisation that the user is part of.
}
\details{
The user can obtain the information of a number of models that have been uploaded on Jaqpot and belong to a specific organisation.
Notethat the user should be part of the organisation .
}
\examples{
 \dontrun{
#get_org_models <- get.orgs.models()
}

}
