% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login.api.unmask}
\alias{login.api.unmask}
\title{Jaqpot login using api key}
\usage{
login.api.unmask(API_key, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The function invisibly returns the jaqpot token, which is also written in the global environment
so that it can be used by other jaqpotr functions.
}
\description{
Provides the user the ability to communicate with the jaqpot server
}
\details{
The user  generates a 
token that is used by jaqpotr functions by providing the jaqpot api key.
}
\examples{
 \dontrun{
login.api.unmask("my_API_key")
}

}
