% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_caret.R
\name{deploy.caret}
\alias{deploy.caret}
\title{Deploy Decision Tree Models on Jaqpot}
\usage{
deploy.caret(
  trained.model,
  preprocess.model = NULL,
  ensemble.model = NULL,
  replace = NULL,
  ymax = NULL,
  ymin = NULL,
  url = "https://api.jaqpot.org/jaqpot/",
  ...
)
}
\arguments{
\item{trained.model}{An object of class "tree" (function \code{ctree(()} of package 'party'). 
#' @param ensemble.model The ensemble model, with default value NULL. Only applicable type is stacked ensemble. Note that if an ensemble model is use, 
then the trained.model should be a list of models.}

\item{preprocess.model}{The preprocess model, with default value NULL. Here it is very important NOT to include the response into the preprocessing phase!
Allowed types of preprocessing: 'preProcess' and 'dummyVars' caret functions.}

\item{replace}{used for NA substitution with a desired value. It should be a list of two arguments, with the first being either "before" or "after", 
for doing the substitution before or after preprocessing, and the second being the desired replacement value.}

\item{ymax}{Maximum y value to be used for y detrasformation. This applies if y has been scaled using the \code{range} method of \code{preProcess()} function.}

\item{ymin}{Minimum y value to be used for y detrasformation. This applies if y has been scaled using the \code{range} method of \code{preProcess()} function.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}

\item{...}{Extra arguments to be passed down the R client. This is not recommended.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained decision tree models on Jaqpot given
a "tree" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the
 function \code{ctree()} of package 'party'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #party.model <- party::ctree(y~x, data=df)
 #deploy.party(tree.model)
}

}
