% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_gbm.R
\name{deploy.gbm}
\alias{deploy.gbm}
\title{Deploy Generalized Boosted Regression Models (GBM) on Jaqpot}
\usage{
deploy.gbm(object, replace = NULL, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{object}{An object of class "gbm" (function \code{gbm()} of package 'gbm').
@param replace used for NA substitution with a desired numeric value.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads a trained GBM model on Jaqpot given a "gbm" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the
  \code{gbm()} function of package 'gbm'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #gbm.model <- gbm::gbm(y~x, data=df)
 #deploy.gbm(gbm.model)
 }

}
