% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_randomForest.R
\name{deploy.randomForest}
\alias{deploy.randomForest}
\title{Deploy Random Forest Models on Jaqpot}
\usage{
deploy.randomForest(object, url = "https://api.jaqpot.org/")
}
\arguments{
\item{object}{An object of class "randomForest.formula"
(function \code{randomForest()} of package 'randomForest').}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained random forest models on Jaqpot given
a "randomForest.formula" object (function \code{(randomForest)} 
of package 'randomForest').
}
\details{
The user can upload on Jaqpot a model that has been trained using the
 function function \code{(randomForest)} 
of package 'randomForest'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #rf.model <- randomForest::randomForest(y~x, data=df)
 #deploy.randomForest(rf.model)
 }

}
