% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_svm.R
\name{deploy.svm}
\alias{deploy.svm}
\title{Deploy Support Vector Machine (SVM) Models on Jaqpot}
\usage{
deploy.svm(object, replace = NULL, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{object}{An object of  class "svm.formula" (function \code{svm()} of package 'e1071').}

\item{replace}{used for NA substitution with a desired numeric value.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained SVM models on Jaqpot given
an "svm.formula" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the function
  \code{svm()} of package 'e1071'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #svm.model <- e1071::svm(y~x, data=df)
 #deploy.svm(tree.model)
 }

}
