% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_trees.R
\name{deploy.tree}
\alias{deploy.tree}
\title{Deploy Tree Models on Jaqpot}
\usage{
deploy.tree(object, replace = NULL, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{object}{An object of class "tree" (function \code{tree()} of package 'tree').}

\item{replace}{used for NA substitution with a desired numeric value.}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
The id of the uploaded model.
}
\description{
Uploads trained tree models on Jaqpot given
a "tree" object.
}
\details{
The user can upload on Jaqpot a model that has been trained using the
 function \code{tree()} of package 'tree'. Apart from the model object, the user is requested
 to provide further information (i.e. Jaqpot API key or credentials, model title and short
 description etc.) via prompt messages. If the upload process is successful,
 the user is given a unique model id key.
}
\examples{
 \dontrun{
 #tree.model <- tree(y~x, data=df)
 #deploy.tree(tree.model)
}

}
