% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaqpot_predict.R
\name{jaqpot.predict}
\alias{jaqpot.predict}
\title{Obtain Predictions from a Jaqpot Model}
\usage{
jaqpot.predict(df, modelID, url = "https://api.jaqpot.org/jaqpot/")
}
\arguments{
\item{df}{A dataframe containing the prediction dataset. The column names of the dataset
should be identical to the independent features of the model.}

\item{modelID}{The Jaqpot id of the model}

\item{url}{The base path of Jaqpot services. This argument is optional and needs 
to be changed only if an alternative Jaqpot installation is used.}
}
\value{
Model predictions for the provided dataset.
}
\description{
The user can provide a dataset and obtain predictions from an uploaded model given the model id.
}
\details{
The user can obtain predictions from a model that has been uploaded on Jaqpot by providing a dataset 
 and the appropriate model id.
}
\examples{
 \dontrun{
 #df <- data.frame(x1 = c(1,2), x2 = c(0.5,0.5))
 #df$x3 <- list(c(1,1), c(2,2))
 #jaqpot.predict(df = df, modelID = "Lfz3aBdh4LlJIJxAvNqV")
}

}
